fix: workers deps + misc refactor
