+++ tsconfig.json
@@ -3,7 +3,7 @@
     "src/index.ts",
     "src/processing/world_compute_worker.ts",
     "src/node/world_compute_node_worker.ts",
-    "src/tests/chunks_over_ws_service.ts"
+    "src/node-tools/chunks_over_ws_server.ts"
   ],
   "compilerOptions": {
     /* Type Checking */
